+++ dbdaora/geospatial/_tests/test_integration_service_geospatial_aioredis_add.py
@@ -1,6 +1,8 @@
 import asynctest
 import pytest
 from aioredis import RedisError
+from circuitbreaker import CircuitBreakerError
+from pymongo.errors import PyMongoError
 
 
 @pytest.mark.asyncio
@@ -40,3 +42,16 @@ async def test_should_add_to_fallback_after_open_circuit_breaker(
 
     assert entity == fake_entity
     assert fake_service.logger.warning.call_count == 2
+
+
+@pytest.mark.asyncio
+async def test_should_add_to_fallback_after_open_fallback_circuit_breaker(
+    fake_service, fake_entity, fake_entity_add, fake_entity_add2
+):
+    fake_put = asynctest.CoroutineMock(side_effect=PyMongoError)
+    fake_service.repository.fallback_data_source.put = fake_put
+
+    with pytest.raises(CircuitBreakerError):
+        await fake_service.add(fake_entity_add, memory=False)
+
+    assert fake_service.logger.warning.call_count == 1